refactor: game 패키지 이름을 evaluation으로 변경
