+++ src/main/java/chess/model/board/Turn.java
@@ -1,4 +1,4 @@
-package chess.model.game;
+package chess.model.board;
 
 import chess.model.piece.Piece;
 import chess.model.piece.Side;
