+++ src/main/java/chess/model/piece/Piece.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.game.PieceValue;
+import chess.model.evaluation.PieceValue;
 import chess.model.position.Position;
 import chess.model.position.Path;
