+++ src/test/java/chess/model/evaluation/PositionEvaluationTest.java
@@ -1,4 +1,4 @@
-package chess.model.game;
+package chess.model.evaluation;
 
 import chess.model.board.ChessBoard;
 import chess.model.board.ChessBoardInitializer;
